Use system libjuce
The submodule is still present for the MacOS build, but no longer used
on Debian.  JUCE is already packaged for Debian, and so we can use the
version from there, which allows us to stay up to date, etc.

Also, add a flag to the debianization to only link libraries we actually
need, since the Projucer adds flags for ones we're not using and that
clutters up the dependencies.

Update build-dependencies and the README list based on packages actually
found to be needed when building on a clean Debian sid.